+++ library/src/main/java/eu/kanade/tachiyomi/network/ProgressListener.kt
@@ -1,0 +1,9 @@
+@file:Suppress("Unused", "UnusedReceiverParameter")
+
+package eu.kanade.tachiyomi.network
+
+fun interface ProgressListener {
+    fun onProgress(percentage: Float)
+}
+
+fun ProgressListener.onProgress(bytesRead: Long, contentLength: Long): Unit = throw Exception("Stub!")